% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{argmax}
\alias{argmax}
\title{Calculate argmax function}
\usage{
argmax(vec)
}
\arguments{
\item{vec}{Input R vector}
}
\value{
The maximum index with the maximum value of the given vector
}
\description{
Calculate argmax function
}
