% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{argmin}
\alias{argmin}
\title{Calculate argmin function}
\usage{
argmin(vec)
}
\arguments{
\item{vec}{Input R vector}
}
\value{
The minimum index with the minimum value of the given vector
}
\description{
Calculate argmin function
}
