% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{cantor_mapping}
\alias{cantor_mapping}
\title{Function to implement cantor mapping for two ordered pair of numbers}
\usage{
cantor_mapping(k1, k2)
}
\arguments{
\item{k1}{First input}

\item{k2}{Second input}
}
\value{
Cantor mapping value
}
\description{
Function to implement cantor mapping for two ordered pair of numbers
}
