% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{covert_to_binary}
\alias{covert_to_binary}
\title{Change the class label to {+1} and {-1} on pairwise classes}
\usage{
covert_to_binary(class1, class2, data.dim, seed)
}
\arguments{
\item{class1}{First pairwise class}

\item{class2}{Second pairwise class}

\item{data.dim}{Data dimensionality}

\item{seed}{Set random seed}
}
\value{
Pairwise classes with correct class label as {+1} and {-1}
}
\description{
Change the class label to {+1} and {-1} on pairwise classes
}
