% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{data_one2all}
\alias{data_one2all}
\title{Function to generate K-size OVA data}
\usage{
data_one2all(dataXY, data.dim, num_class)
}
\arguments{
\item{dataXY}{Full multiclass data set}

\item{data.dim}{Data dimesnionality}

\item{num_class}{Total number of classes, as K}
}
\value{
A list include K OVA data
}
\description{
Function to generate K-size OVA data
}
