% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{dmnorm}
\alias{dmnorm}
\title{Calculate the density of multivariate normal}
\usage{
dmnorm(x_mu, Sigma)
}
\arguments{
\item{x_mu}{Mean vector}

\item{Sigma}{Covariance matrix}
}
\value{
Density of multivariate normal distribution
}
\description{
Calculate the density of multivariate normal
}
