% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.r
\name{evaluate_performance}
\alias{evaluate_performance}
\title{Function to evaluate the multi-class SVM probability estimation performance}
\usage{
evaluate_performance(multiclass.prob, methods = list(type = "pairwise"))
}
\arguments{
\item{multiclass.prob}{Multiclass probability estimation matrix}

\item{methods}{Multiclass probability estimation methods: e.g: "pairwise", "one2rest","liner_time"}
}
\value{
Performance evaluation matrix and classiifcation error matrix
}
\description{
Function to evaluate the multi-class SVM probability estimation performance
}
