% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{find.biggest.class}
\alias{find.biggest.class}
\title{Function to find the class with most abundance data points}
\usage{
find.biggest.class(dataXY, data.dim)
}
\arguments{
\item{dataXY}{Full multiclass data set}

\item{data.dim}{Data dimesnionality}
}
\value{
Class label with the biggest class
}
\description{
Function to find the class with most abundance data points
}
