% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{find_median_index}
\alias{find_median_index}
\title{Function to find the index of median value}
\usage{
find_median_index(x)
}
\arguments{
\item{x}{Input vector}
}
\value{
The index of median value
}
\description{
Function to find the index of median value
}
