% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.r
\name{generate_MP}
\alias{generate_MP}
\title{Generate the MP matrix for reliability diagram}
\usage{
generate_MP(mp, n_class, n_data, method)
}
\arguments{
\item{mp}{Multiclass probability matrix}

\item{n_class}{Number of classes}

\item{n_data}{Number of test data}

\item{method}{Multilcass probability estimation method}
}
\value{
Probability matrix for relability diagram
}
\description{
Generate the MP matrix for reliability diagram
}
