% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{gkl.cal}
\alias{gkl.cal}
\title{Calculate the GKL}
\usage{
gkl.cal(p, phat)
}
\arguments{
\item{p}{True probability}

\item{phat}{Estimated probability}
}
\value{
GKL value
}
\description{
Calculate the GKL
}
