% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.dist.r
\name{median_class_distance}
\alias{median_class_distance}
\title{Calculate the mediam of between class distance}
\usage{
median_class_distance(list_classes_X, calculate_between_class_dist = FALSE)
}
\arguments{
\item{list_classes_X}{List of X for all classes}

\item{calculate_between_class_dist}{Boolean to indicate whether to calculate the between class distance for baseline method 2}
}
\value{
Theclass index of the median of between class distance or the SigmaM
}
\description{
Calculate the mediam of between class distance
}
