% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{multiclass.separation}
\alias{multiclass.separation}
\title{Function to separate the data based on Y labels}
\usage{
multiclass.separation(dataXY, data.dim)
}
\arguments{
\item{dataXY}{Full multiclass data set}

\item{data.dim}{Data dimesnionality}
}
\value{
Two lists, one includes y and other any includes X for each class
}
\description{
Function to separate the data based on Y labels
}
