% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass.svm.probability.linear.algorithm.r
\name{multiclass.svm.probability.linear.algorithm}
\alias{multiclass.svm.probability.linear.algorithm}
\title{Calculate the multiclass probability based on the baseline learning}
\usage{
multiclass.svm.probability.linear.algorithm(
  pairwise.prob,
  base_class,
  k_class,
  actual_labels,
  actual_prob,
  pair_indexes
)
}
\arguments{
\item{pairwise.prob}{Pairwise probability matrix from baseline learning}

\item{base_class}{Baseline class}

\item{k_class}{Number of classes}

\item{actual_labels}{True class labels}

\item{actual_prob}{True probability}

\item{pair_indexes}{Pairwise class indexes}
}
\value{
Multiclass class probaility estimation matrix
}
\description{
Calculate the multiclass probability based on the baseline learning
}
