% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass.svm.probability.r
\name{multiclass.svm.probability}
\alias{multiclass.svm.probability}
\title{Calculate the multiclass probability based on the pairwise probability matrix}
\usage{
multiclass.svm.probability(
  pairwise.prob,
  k_class,
  actual_labels,
  actual_prob,
  pair_indexes
)
}
\arguments{
\item{pairwise.prob}{Pairwise probability matrix}

\item{k_class}{Number of classes}

\item{actual_labels}{True class labels}

\item{actual_prob}{True probability}

\item{pair_indexes}{Pairwise class indexes}
}
\value{
Multiclass class probaility estimation matrix
}
\description{
Calculate the multiclass probability based on the pairwise probability matrix
}
