% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{one2all_label_trans}
\alias{one2all_label_trans}
\title{Function to format data label for one to all probability estimation}
\usage{
one2all_label_trans(dataXY, data.dim, sel_class)
}
\arguments{
\item{dataXY}{Full multiclass data set}

\item{data.dim}{Data dimesnionality}

\item{sel_class}{Select class as {+1} class in binary wSVMs}
}
\value{
Data set with two classes for OVA method
}
\description{
Function to format data label for one to all probability estimation
}
