% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{p_normalize}
\alias{p_normalize}
\title{Normalize the probability vector}
\usage{
p_normalize(c)
}
\arguments{
\item{c}{Probability vector}
}
\value{
Normalized probability vector
}
\description{
Normalize the probability vector
}
