% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.svm.probability.linear.algorithm.r
\name{pairwise.svm.probability.linear.algorithm}
\alias{pairwise.svm.probability.linear.algorithm}
\title{Function to perform the baseline learning of multiclass probability estimation}
\usage{
pairwise.svm.probability.linear.algorithm(
  train.data,
  tune.data,
  test.data,
  data_dim,
  kernel = list(type = "linear", param1 = NULL, param2 = NULL),
  linear.time.algorithm = list(type = "largestClSize"),
  tuning.criteria = "GKL",
  seed = 1,
  eps = 1e-10
)
}
\arguments{
\item{train.data}{Training data as list object}

\item{tune.data}{Tuning data as list object}

\item{test.data}{Test data as list object}

\item{data_dim}{Data dimensionality}

\item{kernel}{Kernel to use in wSVMs}

\item{linear.time.algorithm}{Set baseline learning method: 'largestClSize' for method 1; 'medianClDist' for method 2}

\item{tuning.criteria}{Tuning by GKL or EGKL}

\item{seed}{Set random seed}

\item{eps}{Small threshold value}
}
\value{
Baseline learning pairwise probability estimation as a list object
}
\description{
Function to perform the baseline learning of multiclass probability estimation
}
