% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/pairwise.svm.probability.linear.time.reconstruct.r
\name{pairwise.svm.probability.linear.time.reconstruct}
\alias{pairwise.svm.probability.linear.time.reconstruct}
\title{Reconstruct the pairwise conditional probability table based on baseline learning}
\usage{
pairwise.svm.probability.linear.time.reconstruct(
  estimate_prob_binary_classes_base,
  base_class,
  num_class,
  pair_indexes,
  sum_prob_difference,
  actual_labels,
  actual_prob
)
}
\arguments{
\item{estimate_prob_binary_classes_base}{Pairwise probability with baseline learning}

\item{base_class}{Baseline class}

\item{num_class}{Number of classes}

\item{pair_indexes}{Index of pairwise classes}

\item{sum_prob_difference}{Sum of pairwise probability difference}

\item{actual_labels}{True labels}

\item{actual_prob}{True probability}
}
\value{
Reconstructed pairwise coupling probability estimation table from baseline learning as a list object
}
\description{
Reconstruct the pairwise conditional probability table based on baseline learning
}
