% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.svm.probability.r
\name{pairwise.svm.probability}
\alias{pairwise.svm.probability}
\title{Function to estimate the pairwise coupling probability with dynamic programming for multiclass problem}
\usage{
pairwise.svm.probability(
  train.data,
  tune.data,
  test.data,
  data_dim,
  kernel = list(type = "linear", param1 = NULL, param2 = NULL),
  tuning.criteria = "GKL",
  seed = 1,
  eps = 1e-10
)
}
\arguments{
\item{train.data}{Training data as list object}

\item{tune.data}{Tuning data as list object}

\item{test.data}{Test data as list object}

\item{data_dim}{Data dimensionality}

\item{kernel}{Kernel to use in wSVMs}

\item{tuning.criteria}{Tuning by GKL or EGKL}

\item{seed}{Set random seed}

\item{eps}{Small threshold value}
}
\value{
Pairwise coupling probability estimation as a list object
}
\description{
Function to estimate the pairwise coupling probability with dynamic programming for multiclass problem
}
