% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{pairwise_prob_infer}
\alias{pairwise_prob_infer}
\title{Function to reconstruct the pairwise probability table from baseline learning}
\usage{
pairwise_prob_infer(pjb, pjpb)
}
\arguments{
\item{pjb}{Pairwise conditional probability of pj given (j,base)}

\item{pjpb}{Pairwise conditional probability of pj' given (j',base)}
}
\value{
Pairwise conditional probability of pj given (j,j')
}
\description{
Function to reconstruct the pairwise probability table from baseline learning
}
