% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.estimate.r
\name{prob.estimate}
\alias{prob.estimate}
\title{Estimate the binary class probability}
\usage{
prob.estimate(v, p)
}
\arguments{
\item{v}{Predicted Labels}

\item{p}{Vector of weights}
}
\value{
Estimated probability for {+1} class
}
\description{
Estimate the binary class probability
}
