% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.r
\name{prob.evaluation}
\alias{prob.evaluation}
\title{Evaluate the probability estimation performance}
\usage{
prob.evaluation(p, phat, method = list(type = "egkl"))
}
\arguments{
\item{p}{True probability matrix}

\item{phat}{Estimated probability matrix}

\item{method}{Evaluation methods}
}
\value{
Evaluation performance value
}
\description{
Evaluate the probability estimation performance
}
