% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QP.compact.matrix.r
\name{QP.compact.matrix}
\alias{QP.compact.matrix}
\title{Compute the compact matrix representation}
\usage{
QP.compact.matrix(Amat)
}
\arguments{
\item{Amat}{Input Matrix}
}
\value{
Compact matrix form for quodprog input
}
\description{
Compute the compact matrix representation
}
