% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.dist.r
\name{within_class_compactness}
\alias{within_class_compactness}
\title{Function to calculate the within class compactness}
\usage{
within_class_compactness(classX)
}
\arguments{
\item{classX}{Input class for calculate the Dcp}
}
\value{
Within class compactness
}
\description{
Function to calculate the within class compactness
}
