% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.funs.r
\name{wsvm.egkl}
\alias{wsvm.egkl}
\title{Calculate the EGKL}
\usage{
wsvm.egkl(y, px)
}
\arguments{
\item{y}{Label vector}

\item{px}{Estimated probability}
}
\value{
EGKL value
}
\description{
Calculate the EGKL
}
