% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsvm.kernel.r
\name{wsvm.kernel}
\alias{wsvm.kernel}
\title{Function to calculate the Kernel matrix}
\usage{
wsvm.kernel(U, V, kernel = list(type = "linear", param1 = NULL, param2 = NULL))
}
\arguments{
\item{U}{U matrix}

\item{V}{V matrix}

\item{kernel}{Kernel params list.
kernel$type = type of kernel,
(eg: 'linear', 'poly', 'rbf', 'nn').
kernel$param = parameter of kernel,
(eg: param1 = d degree for 'poly',
param1 = sigma scale for 'rbf',
param1,2 = kapa1 and kapa2 for 'nn')}
}
\value{
A nrow(U) x nrow(V) kernel matrix
}
\description{
Function to calculate the Kernel matrix
}
