% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsvm.prob.r
\name{wsvm.prob}
\alias{wsvm.prob}
\title{Class probability estimation for giving data}
\usage{
wsvm.prob(
  x.train,
  y.train,
  x.test,
  y.test,
  kernel = list(type = "linear", param1 = NULL, param2 = NULL),
  lambda,
  eps = 1e-10
)
}
\arguments{
\item{x.train}{Training data in matrix}

\item{y.train}{Training labels}

\item{x.test}{Test data in matrix}

\item{y.test}{Test labels}

\item{kernel}{Kernel to use for wSVMs}

\item{lambda}{Regularization params}

\item{eps}{small threshold}
}
\value{
List of probability estimation object
}
\description{
Class probability estimation for giving data
}
