% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsvm.r
\name{wsvm}
\alias{wsvm}
\title{Fit an weighted svm model with given weight pi on negative class data}
\usage{
wsvm(
  x.train,
  y.train,
  PI = 0.5,
  kernel = list(type = "linear", param1 = NULL, param2 = NULL),
  lambda = 0.01,
  eps = 1e-10
)
}
\arguments{
\item{x.train}{Training data in matrix form}

\item{y.train}{Label as vector}

\item{PI}{Weight on negative class data}

\item{kernel}{Kernel to choose, refer to function wsvm.kernel}

\item{lambda}{Regulatory parameter for solving weighted SVMs, lambda > 0}

\item{eps}{Small threshold}
}
\value{
Model object as a list consists of model fit: c, d, alpha, and sv
}
\description{
Fit an weighted svm model with given weight pi on negative class data
}
